Add write enable and UUID display
Programmed actual EPDF boards with hat eeprom, discovered a few
small improvements needed. I was able to swap hats and rerun
the script without a header on the hat. Lift up a little bit on
the far side of the hat to make good steady electrical contact
and then run epdf_flash.sh and wait for success and UUID display.